Add getUserWithCandidate method to UserRepository
Introduced a new method called getUserWithCandidate in UserRepository that retrieves a user along with their social media and candidate information. This method uses a combined SQL statement and PHP's PDO to query the user with given id. Providing a means to fetch a user's collective data from multiple related tables in one go.