Implement avatar upload feature in CandidateController
A new function, uploadAvatar(), has been added to the CandidateController. This function allows candidates to upload an avatar, throws appropriate errors if the candidate doesn't exist or if there are server issues, and returns successful upload messages. A POST route to this function in the routes.yaml file should be created for candidates to use this feature.