Refactor updateJobOffer function in JobOfferService
The type hint of updateJobOffer function in JobOfferService.php file has been modified. This function now explicitly demands a JobOffer type object, hence simplifying the code by eliminating the need to check if the argument is an instance of JobOffer within the function body. This makes the function easier to read and maintain.