Refactor RequestValidatorService class
The RequestValidatorService class was moved to the root directory of the Services folder and significant changes were made to improve data validation. The main change was the addition of a function to check for required fields in the object. This not only aids accurate data validation but also helps to return more informative debug messages.