Update FileManagerService functions documentation
The documentation for the functions in the FileManagerService class have been updated to be more informative. The changes include addition of detailed descriptions for the parameters and the returned values, as well as clearer explanations on when exceptions are thrown. This will allow easier understanding of what each function does, and how to use them.