Replace BindValueService with TransactionDbService in ApplyRepository
This change updates the use of the deprecated 'bindValueService' with the newer 'transactionDbService' in the 'ApplyRepository' class. It also simplifies the class by directly using statement bindValue method instead of going through an additional service. Lastly, adjustments were made in the insert, update, delete and select statements to leverage the functionality provided by the 'transactionDbService'.